% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceaccounted-package.R
\docType{package}
\name{forceaccounted}
\alias{forceaccounted}
\alias{forceaccounted-package}
\title{Data for The Force Accounted}
\description{
\url{http://www.bloomberg.com/graphics/2015-star-wars-the-force-accounted/}
}
\author{
Bob Rudis (@hrbrmstr)
}

